data-ingestion/data-ingestion-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pymongo==4.10.1
google-cloud-storage==2.18.2
google-auth==2.35.0
pydantic==2.9.2
